Display markup delimiters properly
Testing with the example program `square.typo` revealed a display bug:
since the working string was written to the display's inner HTML,
putting HTML markup delimiters like < and & in the working string could
break the display.

To fix this bug, we upgrade from writing HTML to manipulating the DOM.
This takes a few more lines of code, but it's safer, cleaner, and more
efficient (because it doesn't require HTML parsing).

Now that we can store highlight colors directly as span elements, it's
a good time to split the resource dictionaries off into a ResourceKit
class.