Apply Monorail style to application page
To keep `monorail.css` from messing up our buttons, we restrict the
wildcard style to `body`. (Only the `padding: 0` is causing problems,
but it seems prudent to restrict everything.)

The wildcard style is presumably doing something useful on the main
site (school/web), so we'll have to test this patch there and revise it
as needed.