Go back to only saving loaded code for comparison
This reverts part of commit b009ffb. If the user gets an error, they
don't need to reload their previous error-free code: it's still loaded!
To remind them of this, we give the disabled styling precedence over the
error styling, so the load button is visibly disabled when the currently
loaded code is restored.

Also, improve class name for pressed buttons.